@@ -34,14 +34,17 @@
     "@typescript-eslint/parser": "4.29.3",
     "babel-jest": "27.3.1",
     "babel-plugin-module-resolver": "4.1.0",
-    "eslint": "7.12.1",
+    "eslint": "7.32.0",
     "eslint-config-airbnb-typescript": "14.0.1",
+    "eslint-config-prettier": "8.3.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-prettier": "4.0.0",
     "eslint-plugin-promise": "4.2.1",
     "factory-girl": "5.0.4",
     "faker": "5.5.3",
     "jest": "27.3.1",
+    "prettier": "2.4.1",
     "supertest": "6.1.6",
     "ts-jest": "27.0.7",
     "ts-node-dev": "1.1.8",
