@@ -40,16 +40,24 @@ describe('src :: api :: services :: authenticate', () => {
   it('should throw invalid value error when trying to authenticate', async () => {
     const temp = await factory.create<Person>('People');
     try {
-      const result = await AuthenticateService.authenticate(temp.email, '123456');
+      const result = await AuthenticateService.authenticate(
+        temp.email,
+        '123456'
+      );
     } catch (e) {
       expect(e).toBeInstanceOf(InvalidValue);
-      expect((<InvalidValue>e).message).toBe('The value ****** for senha is invalid');
+      expect((<InvalidValue>e).message).toBe(
+        'The value ****** for senha is invalid'
+      );
     }
   });
 
   it('should throw invalid value error when trying to authenticate', async () => {
     try {
-      const result = await AuthenticateService.authenticate('Joazinho@mail.com', '123456');
+      const result = await AuthenticateService.authenticate(
+        'Joazinho@mail.com',
+        '123456'
+      );
     } catch (e) {
       expect(e).toBeInstanceOf(NotFound);
       expect((<NotFound>e).message).toBe('Value Joazinho@mail.com not found');
