@@ -1,7 +1,7 @@
-declare namespace Express{
+declare namespace Express {
   export interface Request {
     userId: string;
     habilitado: string;
-    email:string;
+    email: string;
   }
 }
