@@ -4,7 +4,7 @@ import CarService from '@services/CarService';
 import { Request, Response, NextFunction } from 'express';
 
 class CarController {
-  async create(req:Request, res:Response, next: NextFunction) {
+  async create(req: Request, res: Response, next: NextFunction) {
     try {
       const car = await CarService.create(req.body);
       return res.status(201).json(car);
@@ -13,7 +13,7 @@ class CarController {
     }
   }
 
-  async get(req:Request, res:Response, next: NextFunction) {
+  async get(req: Request, res: Response, next: NextFunction) {
     try {
       const cars = await CarService.list(req.query as CarSearch);
       return res.status(200).json(cars);
@@ -22,7 +22,7 @@ class CarController {
     }
   }
 
-  async getById(req:Request, res:Response, next: NextFunction) {
+  async getById(req: Request, res: Response, next: NextFunction) {
     try {
       const { id } = req.params;
       const car = await CarService.getById(id);
@@ -32,27 +32,41 @@ class CarController {
     }
   }
 
-  async update(req:Request, res:Response, next: NextFunction) {
+  async update(req: Request, res: Response, next: NextFunction) {
     try {
       const { id } = req.params;
       const updated = await CarService.update(id, req.body);
       if (updated) {
         return res.status(200).json(updated);
       }
-      return res.status(400).send({ message: 'Something went wrong!' });
+      return res
+        .status(400)
+        .send([{ description: 'Bad Request', name: 'Something went wrong!' }]);
     } catch (e) {
       return next(e);
     }
   }
 
-  async delete(req:Request, res:Response, next: NextFunction) {
+  async delete(req: Request, res: Response, next: NextFunction) {
     try {
       const { id } = req.params;
       const removed = await CarService.delete(id);
       if (removed) {
         return res.status(204).end();
       }
-      return res.status(400).send({ message: 'Something went wrong!' });
+      return res
+        .status(400)
+        .send([{ description: 'Bad Request', name: 'Something went wrong!' }]);
+    } catch (e) {
+      return next(e);
+    }
+  }
+
+  patchAcessorios(req: Request, res: Response, next: NextFunction) {
+    try {
+      const { id, idAcessorios } = req.params;
+      const updated = CarService.updateAccessory(id, idAcessorios, req.body);
+      return res.status(200).json(updated);
     } catch (e) {
       return next(e);
     }
