@@ -3,7 +3,7 @@ import PeopleService from '@services/PeopleService';
 import { Request, Response, NextFunction } from 'express';
 
 class PeopleController {
-  async create(req:Request, res:Response, next: NextFunction) {
+  async create(req: Request, res: Response, next: NextFunction) {
     try {
       const people = await PeopleService.create(req.body);
 
@@ -13,7 +13,7 @@ class PeopleController {
     }
   }
 
-  async get(req:Request, res:Response, next: NextFunction) {
+  async get(req: Request, res: Response, next: NextFunction) {
     try {
       const people = await PeopleService.list(req.query);
       return res.status(200).json(people);
@@ -22,7 +22,7 @@ class PeopleController {
     }
   }
 
-  async getById(req:Request, res:Response, next: NextFunction) {
+  async getById(req: Request, res: Response, next: NextFunction) {
     try {
       const people = await PeopleService.getById(req.params.id);
       return res.status(200).json(people);
@@ -31,7 +31,7 @@ class PeopleController {
     }
   }
 
-  async update(req:Request, res:Response, next: NextFunction) {
+  async update(req: Request, res: Response, next: NextFunction) {
     try {
       const { id } = req.params;
       const updated = await PeopleService.update(id, req.body);
@@ -44,7 +44,7 @@ class PeopleController {
     }
   }
 
-  async delete(req:Request, res:Response, next: NextFunction) {
+  async delete(req: Request, res: Response, next: NextFunction) {
     try {
       const { id } = req.params;
       const removed = await PeopleService.delete(id);
