@@ -1,9 +1,9 @@
 export default class InvalidValue extends Error {
-  public status:number;
+  public status: number;
 
-  public description:string;
+  public description: string;
 
-  constructor(field:string, value:string, useDefault:boolean = false) {
+  constructor(field: string, value: string, useDefault = false) {
     let message = `The value '${value}' is invalid`;
     if (useDefault) {
       message = value;
