@@ -1,10 +1,10 @@
 import Accessory from './Accessory';
 
 export default interface Car {
-  id?: string
-  modelo: string
-  cor: string
-  ano: Number
-  acessorios: Accessory[]
-  quantidadePassageiros: Number
+  id?: string;
+  modelo: string;
+  cor: string;
+  ano: number;
+  acessorios: Accessory[];
+  quantidadePassageiros: number;
 }
