@@ -1,10 +1,10 @@
 import { Pagination } from './Pagination';
 
 export interface CarSearch extends Pagination {
-  modelo?: string
-  cor?: string
-  ano?: Number
-  descricao?: string
-  quantidadePassageiros?: Number
-  'acessorios.descricao'?: string
+  modelo?: string;
+  cor?: string;
+  ano?: number;
+  descricao?: string;
+  quantidadePassageiros?: number;
+  'acessorios.descricao'?: string;
 }
