@@ -1,5 +1,5 @@
 export interface GenerateTokenContent {
-  id: string
-  email: string
-  habilitado: string
+  id: string;
+  email: string;
+  habilitado: string;
 }
