@@ -1,9 +1,9 @@
 export interface Person {
-  id?: string
-  nome: string
-  cpf: string
-  data_nascimento: string | Date
-  email: string
-  senha: string
-  habilitado: string
+  id?: string;
+  nome: string;
+  cpf: string;
+  data_nascimento: string | Date;
+  email: string;
+  senha: string;
+  habilitado: string;
 }
