@@ -9,15 +9,21 @@ const PersonSchema = new mongoose.Schema({
   data_nascimento: {
     type: Date,
     required: true,
-    transform: (val:Date) => moment(val).format('DD/MM/YYYY'),
+    transform: (val: Date) => moment(val).format('DD/MM/YYYY'),
   },
   email: {
-    type: String, required: true, lowercase: true, unique: true,
+    type: String,
+    required: true,
+    lowercase: true,
+    unique: true,
   },
   senha: { type: String, required: true, select: false },
   habilitado: { type: String, required: true, enum: ['sim', 'não'] },
   dataCriacao: {
-    type: Date, default: Date.now, immutable: true, select: false,
+    type: Date,
+    default: Date.now,
+    immutable: true,
+    select: false,
   },
   dataAtualizacao: { type: Date, default: Date.now, select: false },
   __v: { type: Number, select: false },
