@@ -6,43 +6,52 @@ import PersonSearch from '@interfaces/PersonSearch';
 import { FindUserPayload } from '@interfaces/FindUserPayload';
 
 class PeopleRepository {
-  async create(payload:Person):Promise<Person> {
+  async create(payload: Person): Promise<Person> {
     return PersonModel.create(payload);
   }
 
-  async findAll(payload:PersonSearch, offset:number, limit:number):Promise<PeopleModel> {
+  async findAll(
+    payload: PersonSearch,
+    offset: number,
+    limit: number
+  ): Promise<PeopleModel> {
     const count = await PersonModel.countDocuments(payload);
-    const people = await PersonModel.find(payload, null, { skip: offset * limit, limit }).exec();
+    const people = await PersonModel.find(payload, null, {
+      skip: offset * limit,
+      limit,
+    }).exec();
     const offsets = Math.round(count / limit);
     return new PeopleModel(people, count, limit, offset, offsets);
   }
 
-  async delete(id:string):Promise<boolean> {
+  async delete(id: string): Promise<boolean> {
     await PersonModel.findByIdAndRemove(id).exec();
     return true;
   }
 
-  async findById(id:string):Promise<PersonSearch> {
-    return await PersonModel.findById(id) as PersonSearch;
+  async findById(id: string): Promise<PersonSearch> {
+    return (await PersonModel.findById(id)) as PersonSearch;
   }
 
-  validId(id:string):boolean {
+  validId(id: string): boolean {
     return isValid(id);
   }
 
-  async update(id:string, payload:Person) {
-    return await PersonModel.findByIdAndUpdate(id, payload,
-      {
-        returnOriginal: false,
-      }).exec() as Person;
+  async update(id: string, payload: Person) {
+    return (await PersonModel.findByIdAndUpdate(id, payload, {
+      returnOriginal: false,
+    }).exec()) as Person;
   }
 
-  async findUser(payload:FindUserPayload) {
-    return await PersonModel.findOne(payload,
-      { senha: true, habilitado: true, email: true }).exec() as Person;
+  async findUser(payload: FindUserPayload) {
+    return (await PersonModel.findOne(payload, {
+      senha: true,
+      habilitado: true,
+      email: true,
+    }).exec()) as Person;
   }
 
-  async getUserEmailOrCpf(email:string, cpf:string, id?:string) {
+  async getUserEmailOrCpf(email: string, cpf: string, id?: string) {
     let filter;
     if (id) {
       filter = { $and: [{ $or: [{ email }, { cpf }] }, { _id: { $ne: id } }] };
@@ -50,10 +59,10 @@ class PeopleRepository {
       filter = { $or: [{ email }, { cpf }] };
     }
 
-    return await PersonModel.findOne(
-      filter,
-      { email: true, cpf: true },
-    ).exec() as Person;
+    return (await PersonModel.findOne(filter, {
+      email: true,
+      cpf: true,
+    }).exec()) as Person;
   }
 }
 
