@@ -6,10 +6,8 @@ import Joi from 'joi';
 export default async (req: Request, res: Response, next: NextFunction) => {
   try {
     const schema = Joi.object({
-      email: Joi.string().trim().email()
-        .required(),
-      senha: Joi.string().trim().min(6)
-        .required(),
+      email: Joi.string().trim().email().required(),
+      senha: Joi.string().trim().min(6).required(),
     });
 
     const { error } = schema.validate(req.body, { abortEarly: false });
