@@ -6,10 +6,16 @@ import transformToArray from '@validations/utils/transformJoiResult';
 export default async (req: Request, res: Response, next: NextFunction) => {
   try {
     const schema = Joi.object({
-      id: Joi.string().length(24).trim().regex(/[0-9A-Fa-f]{24}/)
+      id: Joi.string()
+        .length(24)
+        .trim()
+        .regex(/[0-9A-Fa-f]{24}/)
         .message('Invalid Id')
         .required(),
-      idAcessorio: Joi.string().length(24).trim().regex(/[0-9A-Fa-f]{24}/)
+      idAcessorio: Joi.string()
+        .length(24)
+        .trim()
+        .regex(/[0-9A-Fa-f]{24}/)
         .message('Invalid idAcessorio')
         .required(),
     });
