@@ -1,10 +1,10 @@
 import { ErrorPayload } from '@interfaces/ErrorPayload';
 import Joi from 'joi';
 
-const transformToArray = (value:Joi.ValidationError) => {
-  const result:ErrorPayload[] = [];
+const transformToArray = (value: Joi.ValidationError) => {
+  const result: ErrorPayload[] = [];
 
-  value.details.forEach((detail:Joi.ValidationErrorItem) => {
+  value.details.forEach((detail: Joi.ValidationErrorItem) => {
     const name = detail.message;
     const description = detail.path.join('.');
     result.push({ name, description });
