@@ -11,14 +11,14 @@ class MongoDatabase {
         user: config.database.username,
         pass: config.database.password,
         authSource: 'admin',
-      },
+      }
     );
 
     mongoose.connection.on('open', () => {
       console.log('Connected to mongo');
     });
 
-    mongoose.connection.on('error', (err:Error) => {
+    mongoose.connection.on('error', (err: Error) => {
       console.log(err);
     });
   }
