@@ -11,8 +11,18 @@ export default (prefix = '/car'): Router => {
   router.post(`${prefix}/`, PostPutCarValidation, CarController.create);
   router.get(`${prefix}/`, GetCarValidation, CarController.get);
   router.get(`${prefix}/:id`, IdValidation, CarController.getById);
-  router.put(`${prefix}/:id`, IdValidation, PostPutCarValidation, CarController.update);
+  router.put(
+    `${prefix}/:id`,
+    IdValidation,
+    PostPutCarValidation,
+    CarController.update
+  );
   router.delete(`${prefix}/:id`, IdValidation, CarController.delete);
-  router.patch(`${prefix}/:id/acessorios/:idAcessorio`, PatchIdValidationValidation, DescriptionValidation, CarController.patchAcessorios);
+  router.patch(
+    `${prefix}/:id/acessorios/:idAcessorio`,
+    PatchIdValidationValidation,
+    DescriptionValidation,
+    CarController.patchAcessorios
+  );
   return router;
 };
