@@ -9,7 +9,12 @@ export default (prefix = '/people'): Router => {
   router.post(`${prefix}/`, PostPutPeopleValidation, PeopleController.create);
   router.get(`${prefix}/`, GetPeopleValidation, PeopleController.get);
   router.get(`${prefix}/:id`, IdValidation, PeopleController.getById);
-  router.put(`${prefix}/:id`, IdValidation, PostPutPeopleValidation, PeopleController.update);
+  router.put(
+    `${prefix}/:id`,
+    IdValidation,
+    PostPutPeopleValidation,
+    PeopleController.update
+  );
   router.delete(`${prefix}/:id`, IdValidation, PeopleController.delete);
   return router;
 };
